+++ Source/Core/DolphinWX/Config/GeneralConfigPane.h
@@ -11,7 +11,6 @@
 class wxButton;
 class wxCheckBox;
 class wxChoice;
-class wxRadioBox;
 
 class GeneralConfigPane final : public wxPanel
 {
@@ -31,23 +30,17 @@ private:
 
   void OnDualCoreCheckBoxChanged(wxCommandEvent&);
   void OnCheatCheckBoxChanged(wxCommandEvent&);
-  void OnForceNTSCJCheckBoxChanged(wxCommandEvent&);
   void OnThrottlerChoiceChanged(wxCommandEvent&);
-  void OnCPUEngineRadioBoxChanged(wxCommandEvent&);
   void OnAnalyticsCheckBoxChanged(wxCommandEvent&);
   void OnAnalyticsNewIdButtonClick(wxCommandEvent&);
 
   wxArrayString m_throttler_array_string;
-  wxArrayString m_cpu_engine_array_string;
 
   wxCheckBox* m_dual_core_checkbox;
   wxCheckBox* m_cheats_checkbox;
-  wxCheckBox* m_force_ntscj_checkbox;
 
   wxCheckBox* m_analytics_checkbox;
   wxButton* m_analytics_new_id;
 
   wxChoice* m_throttler_choice;
-
-  wxRadioBox* m_cpu_engine_radiobox;
 };
